backend/requirements.txt
fastapi
uvicorn
transformers
torch
python-dotenv
webvtt-py
yt-dlp